Cleanup various commands and ShipyardSystem.cs, moves it to a Triad partial class
