+++ Content.Server/Light/Components/PoweredLightComponent.cs
@@ -6,7 +6,7 @@ using Robust.Shared.Audio;
 using Robust.Shared.Containers;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
-using Content.Server._NF.Shipyard.Systems; // Triad
+using Content.Server._Triad.Shipyard; // Triad
 
 namespace Content.Server.Light.Components
 {
