+++ Content.Shared/_Triad/Shipyard/Save/ShipSaveMessages.cs
@@ -14,30 +14,3 @@ public sealed class SendShipSaveDataClientMessage : EntityEventArgs
         ShipData = shipData;
     }
 }
-
-[Serializable, NetSerializable]
-public sealed class RequestLoadShipMessage : EntityEventArgs
-{
-    public string YamlData { get; }
-
-    public RequestLoadShipMessage(string yamlData)
-    {
-        YamlData = yamlData;
-    }
-}
-
-[Serializable, NetSerializable]
-public sealed class RequestAvailableShipsMessage : EntityEventArgs
-{
-}
-
-[Serializable, NetSerializable]
-public sealed class SendAvailableShipsMessage : EntityEventArgs
-{
-    public List<string> ShipNames { get; }
-
-    public SendAvailableShipsMessage(List<string> shipNames)
-    {
-        ShipNames = shipNames;
-    }
-}
